Have tmux report helpful error on syntax error
Previously, it would just error with "pane not found", since the pane
was deleted when the command exited.

To improve on this, we now set remain-on-exit on, so that the dead pane
is not closed. In Tmux::Session#capture, we can check the pane_dead and
pane_dead_status variables to report a slightly more helpful error.